修改商品导入

+++ gs-biz-support/gs-commodity/gs-commodity-repository/src/main/java/com/gs/lshly/biz/support/commodity/entity/GoodsInfo.java
@@ -217,7 +217,7 @@ public class GoodsInfo extends Model {
     /**
      * in会员积分价格
      */
-    private Double inMemberPointPrice;
+    private BigDecimal inMemberPointPrice;
     /**
      * 出售类型（0普通，1活动）
      */
